Update
Now Nornir is going to create a failed_hosts.yaml file of the failed hosts and is not going to add the failed hosts to the extended_nornir.log file. Is also going to append to the extended_nornir.log file instead of creating a new one every time. The playbook will run the failed_hosts.yaml file as the inventory if it exists if not is going to run the hosts.yaml but if you want to run the hosts.yaml and not the failed_hosts.yaml then rename or delete the failed_hosts.yaml file. The reason for this change is to be able to run the playbook again only on the failed hosts and append the successful ones in the extended_nornir.log to easily keep track of everything.